+++ users.json
@@ -9,12 +9,12 @@
     ]
   },
   {
-    "id": "cfb70bc6-34cc-4fee-b1bc-8a9883b458d7",
-    "username": "Zoro",
-    "age": 30,
+    "id": "1016d15e-a947-4b3e-ada7-1566987f6a63",
+    "username": "KAKUZU",
+    "age": 50,
     "hobbies": [
-      "football",
-      "dancing"
+      "games",
+      "drinks"
     ]
   }
 ]
